Fix failed functional test
The functional test for this tool fails.  Looking at the XML wrapper, it looks like $comments is invalid....it doesn't exist anywhere else in the xml file.  I know that comments can be specified multiple times however there is no way to get the comments fields to appear multiple times in the UI as far as I can tell.  This change allows the comment to be added and still pass.