Add extra environment vars to libxslt
This is needed so that python can find the library. LD_LIBRARY_PATH is needed for runtime loading of the lxml module, the others to build the
lxml module correctly.

Without this lxml will build against the system libraries, so everything would appear to work if the libxslt/libxml2 development files are
installed.